restrict numbers to i16

+++ src/json_to_xml/array_val.rs
@@ -1,5 +1,3 @@
-use std::i32;
-
 use super::{
     models::{ArrayValType, JsonNull, JsonStr, TokenStage, TokenType},
     state::State,
@@ -63,7 +61,7 @@ pub fn array_val_json_number_open_case(
 
             state.update_to_closed_state();
         }
-        _ => match new_num_as_str.parse::<i32>() {
+        _ => match new_num_as_str.parse::<i16>() {
             Ok(_) => {
                 state.update_token_type(TokenType::JsonArray(TokenStage::Content(
                     ArrayValType::JsonNumber(new_num_as_str),
